fix(reward): roc_auc NaN under AMP — fp16 count overflow (#47)
Under Lightning 16-mixed the PatchGAN forward returns fp16 rewards, so
reward_roc_auc built an fp16 `labels` (ones_like/zeros_like preserve
dtype). The counts n_pos*(n_pos+1)/2 (~5e5) and n_pos*n_neg (~1e6)
overflow fp16's 65504 max to inf, and the final
(sum_pos_ranks - inf) / inf = NaN. This is the sanity-check-fine /
full-epoch-NaN signature: small N (n_pos=16) doesn't overflow →
finite 0.559; the full val set (n_pos=1000) overflows → NaN every epoch.

Real euler run: val roc_auc was NaN for all 50 epochs while rewards were
finite fp16 (range [-62, 173], no inf/nan). Build labels in float32 so
counts are dtype-agnostic. Verified fp16 AUC == fp32 AUC (0.8631) on the
real checkpoint; 146 tests pass (+1 regression test).

+++ src/manifold/modules/reward.py
@@ -66,7 +66,14 @@ def reward_roc_auc(reward_winner: Tensor, reward_loser: Tensor) -> Tensor:
     Returns ``0.5`` (neutral) when one class is absent.
     """
     scores = torch.cat([reward_winner.float(), reward_loser.float()])
-    labels = torch.cat([torch.ones_like(reward_winner), torch.zeros_like(reward_loser)])
+    # float32 (not ones_like(reward_*)): under Lightning 16-mixed the rewards are
+    # fp16, and an fp16 ``labels`` makes ``n_pos``/``n_neg`` fp16 — then
+    # ``n_pos*(n_pos+1)/2`` (~5e5) and ``n_pos*n_neg`` (~1e6) overflow fp16's
+    # 65504 max to inf, turning the AUC into (finite - inf) / inf = NaN. Counts
+    # are dtype-agnostic; build them in float32 so fp16 rewards behave like fp32.
+    labels = torch.cat(
+        [torch.ones_like(reward_winner, dtype=torch.float32), torch.zeros_like(reward_loser, dtype=torch.float32)]
+    )
     n_pos = labels.sum()
     n_neg = labels.numel() - n_pos
     if float(n_pos) == 0 or float(n_neg) == 0:
